+design +errors

+++ src/main/java/com/example/photoshoot/repos/PostRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface PostRepo extends CrudRepository<Post, Integer> {
+public interface PostRepo extends CrudRepository<Post, Long> {
     List<Post> findByTag(String tag);
 }
